Adds base GameObject class to reduce code duplication
