+++ App.cpp
@@ -26,7 +26,7 @@ namespace Engine
 		m_lastFrameTime = m_timer->GetElapsedTimeInSeconds();
 
 		m_ship = new Engine::Ship(this);
-		m_asteroid = new Engine::Asteroid(this);
+		m_asteroid = new Engine::Asteroid(this);		
 	}
 
 	App::~App()
@@ -40,7 +40,7 @@ namespace Engine
 		delete m_ship;
 
 		// Removes asteroid
-		delete m_asteroid;
+		delete m_asteroid;		
 	}
 
 	void App::Execute()
@@ -137,7 +137,7 @@ namespace Engine
 		// Update code goes here
 		//
 		m_ship->Update(DESIRED_FRAME_TIME);
-		m_asteroid->Update(DESIRED_FRAME_TIME);
+		m_asteroid->Update(DESIRED_FRAME_TIME);		
 
 		double endTime = m_timer->GetElapsedTimeInSeconds();
 		double nextTimeFrame = startTime + DESIRED_FRAME_TIME;
@@ -160,7 +160,7 @@ namespace Engine
 
 		// Render code goes here
 		m_ship->Render();
-		m_asteroid->Render();
+		m_asteroid->Render();		
 		
 		SDL_GL_SwapWindow(m_mainWindow);
 	}
